requirements.txt
keyboard
pyopengl
glfw
numpy
pillow
pygame
opensimplex